requirements.txt
google-generativeai>=0.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
